move live tracker DO down a level to its own folder
