+++ api/services/neatqueue/neatqueue.ts
@@ -35,7 +35,7 @@ import { create } from "../../embeds/stats/create";
 import { AssociationReason, GamesRetrievable } from "../database/types/discord_associations";
 import { DiscordError } from "../discord/discord-error";
 import { MapsEmbed } from "../../embeds/maps-embed";
-import { isSuccessResponse } from "../../durable-objects/types";
+import { isSuccessResponse } from "../../durable-objects/live-tracker/types";
 import { NeatQueuePlayersEmbed } from "../../embeds/neatqueue/neatqueue-players-embed";
 import type {
   VerifyNeatQueueResponse,
