+++ api/worker-configuration.d.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 // Generated by Wrangler by running `wrangler types` (hash: 603e3a13c3c880505b9f77c5a57ea078)
-// Runtime types generated with workerd@1.20260515.1 2026-02-28 nodejs_compat
+// Runtime types generated with workerd@1.20260521.1 2026-02-28 nodejs_compat
 interface __BaseEnv_Env {
 	APP_DATA: KVNamespace;
 	DB: D1Database;
@@ -11430,11 +11430,11 @@ interface SendEmail {
     send(message: EmailMessage): Promise<EmailSendResult>;
     send(builder: {
         from: string | EmailAddress;
-        to: string | string[];
+        to: string | EmailAddress | (string | EmailAddress)[];
         subject: string;
         replyTo?: string | EmailAddress;
-        cc?: string | string[];
-        bcc?: string | string[];
+        cc?: string | EmailAddress | (string | EmailAddress)[];
+        bcc?: string | EmailAddress | (string | EmailAddress)[];
         headers?: Record<string, string>;
         text?: string;
         html?: string;
